refactor(apollo): rename checkbox component folder
