+++ client/apollo/react/src/Form/Checkbox/checkboxText/CheckboxTextApollo.tsx
@@ -1,7 +1,7 @@
 import "@axa-fr/design-system-apollo-css/dist/Form/Checkbox/CheckboxText/CheckboxTextApollo.scss";
-import { CheckboxTextCommon, CheckboxTextProps } from "./CheckboxTextCommon";
-import { Checkbox } from "../checkbox/CheckboxApollo";
 import { ItemMessage } from "../../ItemMessage/ItemMessageApollo";
+import { Checkbox } from "../Checkbox/CheckboxApollo";
+import { CheckboxTextCommon, CheckboxTextProps } from "./CheckboxTextCommon";
 
 export const CheckboxText = (props: CheckboxTextProps) => (
   <CheckboxTextCommon
