+++ client/apollo/react/src/Form/Checkbox/checkboxText/CheckboxTextLF.tsx
@@ -1,7 +1,7 @@
 import "@axa-fr/design-system-apollo-css/dist/Form/Checkbox/CheckboxText/CheckboxTextLF.scss";
-import { CheckboxTextCommon, CheckboxTextProps } from "./CheckboxTextCommon";
-import { Checkbox } from "../checkbox/CheckboxLf";
 import { ItemMessage } from "../../ItemMessage/ItemMessageLF";
+import { Checkbox } from "../Checkbox/CheckboxLf";
+import { CheckboxTextCommon, CheckboxTextProps } from "./CheckboxTextCommon";
 
 export const CheckboxText = (props: CheckboxTextProps) => (
   <CheckboxTextCommon
